part4: 4.21 bloglist
1. a blog can be deleted only by the user who added the blog

+++ part4/bloglist/controllers/blogs.js
@@ -51,8 +51,28 @@ blogsRouter.post("/", async (request, response) => {
 });
 
 blogsRouter.delete("/:id", async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id);
-  response.status(204).end();
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  // 没有令牌
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token missing or invalid" });
+  }
+
+  const blogToDelete = await Blog.findById(request.params.id);
+
+  // 要删除的不存在
+  if (!blogToDelete) {
+    return response.status(404).json({ error: "not exists" });
+  }
+
+  // 只允许创建该blog的用户删除
+  if (blogToDelete.user.toString() === decodedToken.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id);
+    response.status(204).end();
+  } else {
+    response
+      .status(403)
+      .json({ error: "You do not have permission to delete this blog" });
+  }
 });
 
 blogsRouter.put("/:id", async (request, response) => {
